move core types to core/types
